+++ deeplearning4j-scaleout/spark/dl4j-spark/src/main/java/org/deeplearning4j/spark/api/worker/ExecuteWorkerMultiDataSetFlatMap.java
@@ -85,8 +85,10 @@ class ExecuteWorkerMultiDataSetFlatMapAdapter<R extends TrainingResult>
                 if (stats)
                     s.logNextDataSetBefore();
                 MultiDataSet next = batchedIterator.next();
+
+                // FIXME: int cast
                 if (stats)
-                    s.logNextDataSetAfter(next.getFeatures(0).size(0));
+                    s.logNextDataSetAfter((int) next.getFeatures(0).size(0));
 
                 if (stats) {
                     s.logProcessMinibatchBefore();
