+++ deeplearning4j-scaleout/spark/dl4j-spark/src/main/java/org/deeplearning4j/spark/impl/graph/scoring/ScoreFlatMapFunctionCGDataSet.java
@@ -31,6 +31,7 @@ import org.nd4j.linalg.factory.Nd4j;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import scala.Tuple2;
+import lombok.val;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -83,7 +84,9 @@ class ScoreFlatMapFunctionCGDataSetAdapter
         while (iter.hasNext()) {
             DataSet ds = iter.next();
             double score = network.score(ds, false);
-            int numExamples = ds.getFeatureMatrix().size(0);
+
+            // FIXME: int cast
+            int numExamples = (int) ds.getFeatureMatrix().size(0);
             out.add(new Tuple2<>(numExamples, score * numExamples));
         }
